requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
azure-core==1.35.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
microsoft-agents-activity==0.3.2
microsoft-agents-authentication-msal==0.3.2
microsoft-agents-hosting-aiohttp==0.3.2
microsoft-agents-hosting-core==0.3.2
msal==1.33.0
multidict==6.6.4
propcache==0.3.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
yarl==1.20.1
